requirements.txt
PyQt5==5.13.2
discord.py==1.5.1